backend/pyproject.toml
[project]
name = "taifex-quant-trading-platform-backend"
version = "0.1.0"
description = "FastAPI backend skeleton for the Taifex Quant Trading Platform."
requires-python = ">=3.12"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pydantic-settings",
  "pytest",
  "httpx",
  "ruff",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
